# Add benchmark for route