--- a/src/benchmarks/CMakeLists.txt
+++ b/src/benchmarks/CMakeLists.txt
@@ -30,6 +30,18 @@ target_link_libraries(match-bench
 	${TBB_LIBRARIES}
 	${MAYBE_SHAPEFILE})
 
+add_executable(route-bench
+	EXCLUDE_FROM_ALL
+	route.cpp
+	$<TARGET_OBJECTS:UTIL>)
+
+target_link_libraries(route-bench
+	osrm
+	${BOOST_BASE_LIBRARIES}
+	${CMAKE_THREAD_LIBS_INIT}
+	${TBB_LIBRARIES}
+	${MAYBE_SHAPEFILE})
+
 add_executable(json-render-bench
 	EXCLUDE_FROM_ALL
 	json_render.cpp
@@ -72,5 +84,6 @@ add_custom_target(benchmarks
 	rtree-bench
 	packedvector-bench
 	match-bench
+  route-bench
 	json-render-bench
-    alias-bench)
+  alias-bench)