<!-- BENCHMARK_RESULTS_START -->
## Benchmark Results
| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1135.23<br/>plain u32: 1148.41<br/>aliased double: 1177.7<br/>plain double: 1169.41 | aliased u32: 1131.56<br/>plain u32: 1158.25<br/>aliased double: 1178.46<br/>plain double: 1187.9 |
| json-render | String: 8.50347ms<br/>Stringstream: 11.1367ms<br/>Vector: 7.62734ms | String: 8.49151ms<br/>Stringstream: 11.5754ms<br/>Vector: 7.60116ms |
| match_ch | Default radius: <br/>4.46884ms/req at 82 coordinate<br/>0.0544981ms/coordinate<br/>Radius 5m: <br/>4.47413ms/req at 82 coordinate<br/>0.0545625ms/coordinate<br/>Radius 10m: <br/>15.2224ms/req at 82 coordinate<br/>0.185639ms/coordinate<br/>Radius 15m: <br/>37.0273ms/req at 82 coordinate<br/>0.451552ms/coordinate<br/>Radius 30m: <br/>314.115ms/req at 82 coordinate<br/>3.83067ms/coordinate | Default radius: <br/>4.47584ms/req at 82 coordinate<br/>0.0545834ms/coordinate<br/>Radius 5m: <br/>4.45385ms/req at 82 coordinate<br/>0.0543152ms/coordinate<br/>Radius 10m: <br/>15.1712ms/req at 82 coordinate<br/>0.185015ms/coordinate<br/>Radius 15m: <br/>36.9638ms/req at 82 coordinate<br/>0.450778ms/coordinate<br/>Radius 30m: <br/>314.3ms/req at 82 coordinate<br/>3.83292ms/coordinate |
| match_mld | Default radius: <br/>3.43738ms/req at 82 coordinate<br/>0.0419192ms/coordinate<br/>Radius 5m: <br/>3.341ms/req at 82 coordinate<br/>0.0407439ms/coordinate<br/>Radius 10m: <br/>12.2525ms/req at 82 coordinate<br/>0.149421ms/coordinate<br/>Radius 15m: <br/>31.4506ms/req at 82 coordinate<br/>0.383544ms/coordinate<br/>Radius 30m: <br/>350.259ms/req at 82 coordinate<br/>4.27146ms/coordinate | Default radius: <br/>3.42251ms/req at 82 coordinate<br/>0.041738ms/coordinate<br/>Radius 5m: <br/>3.39757ms/req at 82 coordinate<br/>0.0414338ms/coordinate<br/>Radius 10m: <br/>12.4339ms/req at 82 coordinate<br/>0.151633ms/coordinate<br/>Radius 15m: <br/>32.2096ms/req at 82 coordinate<br/>0.3928ms/coordinate<br/>Radius 30m: <br/>353.05ms/req at 82 coordinate<br/>4.30548ms/coordinate |
| packedvector | random write:<br/>std::vector 11317.4 ms<br/>util::packed_vector 82088.1 ms<br/>slowdown: 7.25326<br/>random read:<br/>std::vector 11113.2 ms<br/>util::packed_vector 33137 ms<br/>slowdown: 2.98177 | random write:<br/>std::vector 11339.4 ms<br/>util::packed_vector 73886.1 ms<br/>slowdown: 6.51586<br/>random read:<br/>std::vector 11110.3 ms<br/>util::packed_vector 29990.5 ms<br/>slowdown: 2.69934 |
| route_ch |  | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>607.904ms<br/>0.607904ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>382.033ms<br/>0.382033ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>767.876ms<br/>0.767876ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>162.814ms<br/>0.162814ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>103.382ms<br/>0.103382ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>145.274ms<br/>0.145274ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>161.632ms<br/>0.161632ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>103.202ms<br/>0.103202ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>145.496ms<br/>0.145496ms/req |
| route_mld |  | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>725.032ms<br/>0.725032ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>464.2ms<br/>0.4642ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>940.778ms<br/>0.940778ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>276.156ms<br/>0.276156ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>165.956ms<br/>0.165956ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>300.712ms<br/>0.300712ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>276.014ms<br/>0.276014ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>166.176ms<br/>0.166176ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>299.902ms<br/>0.299902ms/req |
| rtree | 1 result:<br/>202.292ms ->  0.0202292 ms/query<br/>10 results:<br/>238.514ms ->  0.0238514 ms/query | 1 result:<br/>208.217ms ->  0.0208217 ms/query<br/>10 results:<br/>255.728ms ->  0.0255728 ms/query |
<!-- BENCHMARK_RESULTS_END -->